requirements.txt
python-dotenv
beautifulsoup4
requests
openai
streamlit